improve -add favorite tool- functionality

+++ src/main/java/com/example/Project/repository/ToolRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.Project.model.Tool;
 
-public interface ToolRepository extends JpaRepository<Tool, Long> {
+public interface ToolRepository extends JpaRepository<Tool, Integer> {
     List<Tool> findByIsActiveTrue();
 
     List<Tool> findByNameContainingIgnoreCase(String name);
